requirements.txt
streamlit
pandas
pdfminer.six
pypdf
unidecode
